ux: add tooltip to copy button
- extract CopyButtonWide to its own component
- address PR feedback in test
