+++ src/components/CopyButton.js
@@ -2,8 +2,6 @@ import React from 'react';
 import cn from 'classnames';
 import { LinkButton } from 'indigo-react';
 
-import { CopyButton as BaseCopyButton } from 'components/Buttons';
-
 import useCopiable from 'lib/useCopiable';
 
 const renderDefaultCopyText = didCopy => (didCopy ? 'Copied!' : 'Copy');
@@ -27,22 +25,5 @@ export default function CopyButton({
       {...rest}>
       {children(didCopy)}
     </As>
-  );
-}
-
-export function CopyButtonWide({ text, children, ...rest }) {
-  const [doCopy, didCopy] = useCopiable(text);
-  return (
-    <BaseCopyButton
-      as="span"
-      onClick={doCopy}
-      text={text}
-      className={cn({
-        black: !didCopy,
-        green3: didCopy,
-      })}
-      {...rest}>
-      {children}
-    </BaseCopyButton>
   );
 }
